New RAMSES release, Jan. 2019
Updated freetype to 2.9.
Internal refactoring and improvements.
Stricter rendering tests.
CMake improvements.

Original hash-id: 79bb773ae3b1e9404f0fbb908d2d1103e5a84abe

+++ client/ramses-client/impl/AppearanceImpl.h
@@ -58,8 +58,8 @@ namespace ramses
         status_t getDepthFunction(EDepthFunc& func) const;
         status_t setDepthWrite(EDepthWrite flag);
         status_t getDepthWriteMode(EDepthWrite& mode) const;
-        status_t setStencilFunc(EStencilFunc func, uint32_t ref, uint8_t mask);
-        status_t getStencilFunc(EStencilFunc& func, uint32_t& ref, uint8_t& mask) const;
+        status_t setStencilFunc(EStencilFunc func, uint8_t ref, uint8_t mask);
+        status_t getStencilFunc(EStencilFunc& func, uint8_t& ref, uint8_t& mask) const;
         status_t setStencilOperation(EStencilOperation sfail, EStencilOperation dpfail, EStencilOperation dppass);
         status_t getStencilOperation(EStencilOperation& sfail, EStencilOperation& dpfail, EStencilOperation& dppass) const;
         status_t setCullingMode(ECullMode mode);
