+++ renderer/RendererLib/RendererLib/include/RendererLib/LoggingDevice.h
@@ -42,7 +42,7 @@ namespace ramses_internal
         virtual void cullMode(ECullMode mode) override;
         virtual void depthFunc(EDepthFunc func) override;
         virtual void depthWrite(EDepthWrite flag) override;
-        virtual void stencilFunc(EStencilFunc func, UInt32 ref, UInt8 mask) override;
+        virtual void stencilFunc(EStencilFunc func, UInt8 ref, UInt8 mask) override;
         virtual void stencilOp(EStencilOp sfail, EStencilOp dpfail, EStencilOp dppass) override;
         virtual void drawMode(EDrawMode mode) override;
         virtual void setViewport(UInt32 x, UInt32 y, UInt32 width, UInt32 height) override;
