+++ renderer/RendererLib/RendererTestUtils/include/DeviceMock.h
@@ -58,7 +58,7 @@ namespace ramses_internal
         MOCK_METHOD1(cullMode, void(ECullMode));
         MOCK_METHOD1(depthFunc, void(EDepthFunc));
         MOCK_METHOD1(depthWrite, void(EDepthWrite));
-        MOCK_METHOD3(stencilFunc, void(EStencilFunc, UInt32, UInt8));
+        MOCK_METHOD3(stencilFunc, void(EStencilFunc, UInt8, UInt8));
         MOCK_METHOD3(stencilOp, void(EStencilOp, EStencilOp, EStencilOp));
         MOCK_METHOD1(drawMode, void(EDrawMode));
         MOCK_METHOD4(setViewport, void(UInt32, UInt32, UInt32, UInt32));
